dis.c - fix 32-bit compile (blob_size, size_t, "%ld")
The disassembler used a size_t for the variable that holds the blob
size, even though it can't ever be bigger than 2^32.

There's also a printf statement that prints the blob size with a %ld
specifier (for a long int), which happens to work fine for size_t on
a 64-bit Linux platform.

It doesn't work on a 32-bit platform where size_t is 32-bit.

The current fix is not the right one.  This and many other incorrect
printf format strings for uint32_t et al will be fixed in ReVAX v0.2.