src/uop.c -- source code formatting changes to make some lines shorter
Uop doesn't understand assembler yet, so the VAX instructions to analyze are
specified in C code as literal constant arrays of uint8_t in calls to
analyze().

Some of those lines get a little long.

They get even longer because I want comments that say what the instructions
are.

This leads gedit's print command to either wrap the lines or cut off the end,
even with a 7pt font size (and 6.5pt seems to be synonymous with 7pt -- and
6pt is too small).

This shrink helped a little.