op-dis: kill sprintf-related warning in the generated operand disassembly code
src/op-dis.h uses sprintf() into a buffer, possibly up to 4 times per operand.

Since the buffer it sprintf's into is a fixed-size buffer and the string it
sprintf's into the buffer uses the same fixed-size buffer type, this could
potentially go wrong (if the code was maleficent).

Let's use snprintf() instead.

Why is there an if-statement to protect the snprintf?  Because otherwise
sizeof(str.ret) - idx could go negative (which is a very large positive
number for a size_t).