test-op, build fix when using -std=c99 flag (#define _POSIX_C_SOURCE)
clock_gettime() and related types and constants are not declared by time.h
in strict C99 mode unless the source code specifically asks for it with a
feature test macro.

I have chosen the earliest version that exports them (199309L) instead of
the newest (200809L).  This might need revisiting later.