test-fp -- warning/error fixes
Explicitly link with libm (-lm in the makefile).  Necessary with a couple
of old compilers, otherwise linking test-fp will fail because log10f()
can't be found.

Use a feature test macro, otherwise mrand48() will not be declared when
including stdlib.h.  This is only necessary because of the -std=c99 flag.

Initialize a struct big_int to zero with "{{0}}" instead of just "{}".
I think this is also necessary because of -std=c99.