test-fp -- show/check MPFR version
MPFR is the multi-precision floating-point library ReVAX uses for
converting VAX fp values to/from strings.

The MPFR version should probably be displayed by all programs that
take the --version parameter.

This code checks that the header and library version match.  This
should probably be part of the startup code for all ReVAX programs
that use MPFR.